refactor : 알림엔티티, 매퍼 경로 변경
